updated wrapper logic
